+++ plugins/transport-grpc/src/main/java/org/opensearch/plugin/transport/grpc/proto/request/document/bulk/ActiveShardCountProtoUtils.java
@@ -9,7 +9,6 @@
 package org.opensearch.plugin.transport.grpc.proto.request.document.bulk;
 
 import org.opensearch.action.support.ActiveShardCount;
-import org.opensearch.protobufs.BulkRequest;
 import org.opensearch.protobufs.WaitForActiveShards;
 
 /**
@@ -33,37 +32,25 @@ public class ActiveShardCountProtoUtils {
      * the wait_for_active_shards parameter from the Protocol Buffer request and applies
      * the appropriate ActiveShardCount setting to the OpenSearch bulk request.
      *
-     * @param bulkRequest The OpenSearch bulk request to modify
-     * @param request The Protocol Buffer request containing the active shard count settings
-     * @return The modified OpenSearch bulk request with updated active shard count settings
+     * @param waitForActiveShards The protobuf object containing the active shard count
+     * @return The modified bulk request
      */
-    public static org.opensearch.action.bulk.BulkRequest getActiveShardCount(
-        org.opensearch.action.bulk.BulkRequest bulkRequest,
-        BulkRequest request
-    ) {
-        if (!request.hasWaitForActiveShards()) {
-            return bulkRequest;
-        }
-        WaitForActiveShards waitForActiveShards = request.getWaitForActiveShards();
+    public static ActiveShardCount parseProto(WaitForActiveShards waitForActiveShards) {
+
         switch (waitForActiveShards.getWaitForActiveShardsCase()) {
             case WaitForActiveShards.WaitForActiveShardsCase.WAIT_FOR_ACTIVE_SHARD_OPTIONS:
                 switch (waitForActiveShards.getWaitForActiveShardOptions()) {
                     case WAIT_FOR_ACTIVE_SHARD_OPTIONS_UNSPECIFIED:
                         throw new UnsupportedOperationException("No mapping for WAIT_FOR_ACTIVE_SHARD_OPTIONS_UNSPECIFIED");
                     case WAIT_FOR_ACTIVE_SHARD_OPTIONS_ALL:
-                        bulkRequest.waitForActiveShards(ActiveShardCount.ALL);
-                        break;
+                        return ActiveShardCount.ALL;
                     default:
-                        bulkRequest.waitForActiveShards(ActiveShardCount.DEFAULT);
-                        break;
+                        return ActiveShardCount.DEFAULT;
                 }
-                break;
             case WaitForActiveShards.WaitForActiveShardsCase.INT32_VALUE:
-                bulkRequest.waitForActiveShards(waitForActiveShards.getInt32Value());
-                break;
+                return ActiveShardCount.from(waitForActiveShards.getInt32Value());
             default:
-                throw new UnsupportedOperationException("No mapping for WAIT_FOR_ACTIVE_SHARD_OPTIONS_UNSPECIFIED");
+                return ActiveShardCount.DEFAULT;
         }
-        return bulkRequest;
     }
 }